+++ src/datasources/results/query-handlers/query-steps/QueryStepsDataSource.test.ts
@@ -26,6 +26,7 @@ import { DataSourceBase } from 'core/DataSourceBase';
 
 const mockSteps = Array(1000).fill({ stepId: '1', name: 'Step 1' });
 const mockPaths = Array(1000).fill({ path: 'path1' });
+const queryStepsUrl = '/nitestmonitor/v2/query-steps';
 
 const mockQueryStepsResponse = {
   steps: [
@@ -62,21 +63,12 @@ describe('QueryStepsDataSource', () => {
     [datastore, backendServer, templateSrv] = setupDataSource(QueryStepsDataSource);
 
     backendServer.fetch
-      .calledWith(requestMatching({ url: '/nitestmonitor/v2/query-steps', method: 'POST' }))
+      .calledWith(requestMatching({ url: queryStepsUrl, method: 'POST' }))
       .mockReturnValue(createFetchResponse(mockQueryStepsResponse));
 
     backendServer.fetch
       .calledWith(requestMatching({ url: '/nitestmonitor/v2/query-result-values', method: 'POST' }))
       .mockReturnValue(createFetchResponse(['name1', 'name2']));
-
-    backendServer.fetch
-      .calledWith(requestMatching({ url: '/nitestmonitor/v2/query-paths', method: 'POST' }))
-      .mockReturnValue(
-        createFetchResponse({
-          paths: [{ path: 'path1' }, { path: 'path2' }],
-          continuationToken: null,
-        })
-      );
   });
 
   afterEach(() => {
@@ -94,7 +86,7 @@ describe('QueryStepsDataSource', () => {
     it('should return undefined if API throws unknown status code error', async () => {
       const error = new Error('API failed');
       backendServer.fetch
-        .calledWith(requestMatching({ url: '/nitestmonitor/v2/query-steps', method: 'POST' }))
+        .calledWith(requestMatching({ url: queryStepsUrl, method: 'POST' }))
         .mockImplementationOnce(() => {
           throw error;
         });
@@ -114,7 +106,7 @@ describe('QueryStepsDataSource', () => {
 
     test('should raise an error when API fails', async () => {
       backendServer.fetch
-        .calledWith(requestMatching({ url: '/nitestmonitor/v2/query-steps' }))
+        .calledWith(requestMatching({ url: queryStepsUrl }))
         .mockReturnValue(createFetchError(400));
 
       await expect(datastore.querySteps()).rejects.toThrow(
@@ -126,7 +118,7 @@ describe('QueryStepsDataSource', () => {
       const publishMock = jest.fn();
       (datastore as any).appEvents = { publish: publishMock };
       backendServer.fetch
-        .calledWith(requestMatching({ url: '/nitestmonitor/v2/query-steps' }))
+        .calledWith(requestMatching({ url: queryStepsUrl }))
         .mockReturnValue(createFetchError(400));
 
       await expect(datastore.querySteps()).rejects.toThrow(
@@ -154,7 +146,7 @@ describe('QueryStepsDataSource', () => {
 
     it('should throw not found error when API returns 404 status', async () => {
       backendServer.fetch
-        .calledWith(requestMatching({ url: '/nitestmonitor/v2/query-steps' }))
+        .calledWith(requestMatching({ url: queryStepsUrl }))
         .mockReturnValue(createFetchError(404));
 
       await expect(datastore.querySteps()).rejects.toThrow(
@@ -164,7 +156,7 @@ describe('QueryStepsDataSource', () => {
 
     it('should throw timeOut error when API returns 504 status', async () => {
       backendServer.fetch
-        .calledWith(requestMatching({ url: '/nitestmonitor/v2/query-steps' }))
+        .calledWith(requestMatching({ url: queryStepsUrl }))
         .mockReturnValue(createFetchError(504));
 
       await expect(datastore.querySteps()).rejects.toThrow(
@@ -208,7 +200,7 @@ describe('QueryStepsDataSource', () => {
 
       expect(backendServer.fetch).toHaveBeenCalledWith(
         expect.objectContaining({
-          url: '/nitestmonitor/v2/query-steps',
+          url: queryStepsUrl,
           data: expect.objectContaining({
             orderBy: 'STARTED_AT',
             descending: false,
@@ -228,7 +220,7 @@ describe('QueryStepsDataSource', () => {
 
       expect(backendServer.fetch).toHaveBeenCalledWith(
         expect.objectContaining({
-          url: '/nitestmonitor/v2/query-steps',
+          url: queryStepsUrl,
           data: expect.objectContaining({
             filter: '(startedAt > "${__from:date}" && startedAt < "${__to:date}")',
           }),
@@ -238,7 +230,7 @@ describe('QueryStepsDataSource', () => {
 
     test('should display an empty cell when properties of type objects are returned as empty objects', async () => {
       backendServer.fetch
-        .calledWith(requestMatching({ url: '/nitestmonitor/v2/query-steps', method: 'POST' }))
+        .calledWith(requestMatching({ url: queryStepsUrl, method: 'POST' }))
         .mockReturnValue(
           createFetchResponse({
             steps: [
@@ -266,7 +258,7 @@ describe('QueryStepsDataSource', () => {
 
     test('should display the JSON stringified value when properties of type object are returned as non-empty objects', async () => {
       backendServer.fetch
-        .calledWith(requestMatching({ url: '/nitestmonitor/v2/query-steps', method: 'POST' }))
+        .calledWith(requestMatching({ url: queryStepsUrl, method: 'POST' }))
         .mockReturnValue(
           createFetchResponse({
             steps: [
@@ -296,7 +288,7 @@ describe('QueryStepsDataSource', () => {
 
     test('should display an empty cell when properties of type array are returned as empty array', async () => {
       backendServer.fetch
-        .calledWith(requestMatching({ url: '/nitestmonitor/v2/query-steps', method: 'POST' }))
+        .calledWith(requestMatching({ url: queryStepsUrl, method: 'POST' }))
         .mockReturnValue(
           createFetchResponse({
             steps: [
@@ -322,7 +314,7 @@ describe('QueryStepsDataSource', () => {
 
     test('should display comma separated values when properties of type array are returned as a non-empty array', async () => {
       backendServer.fetch
-        .calledWith(requestMatching({ url: '/nitestmonitor/v2/query-steps', method: 'POST' }))
+        .calledWith(requestMatching({ url: queryStepsUrl, method: 'POST' }))
         .mockReturnValue(
           createFetchResponse({
             steps: [
@@ -362,7 +354,7 @@ describe('QueryStepsDataSource', () => {
 
     test('should show column header with no data when QuerySteps API returns empty array', async () => {
       backendServer.fetch
-        .calledWith(requestMatching({ url: '/nitestmonitor/v2/query-steps', method: 'POST' }))
+        .calledWith(requestMatching({ url: queryStepsUrl, method: 'POST' }))
         .mockReturnValue(
           createFetchResponse({
             steps: [],
@@ -384,7 +376,7 @@ describe('QueryStepsDataSource', () => {
 
     test('should return no data when Query Steps returns error', async () => {
       backendServer.fetch
-        .calledWith(requestMatching({ url: '/nitestmonitor/v2/query-steps' }))
+        .calledWith(requestMatching({ url: queryStepsUrl }))
         .mockReturnValue(createFetchError(400));
 
       const query = buildQuery({
@@ -415,7 +407,7 @@ describe('QueryStepsDataSource', () => {
       jest.spyOn(DataSourceBase.prototype, 'getWorkspaces').mockResolvedValue([]);
       const [datastore, backendServer] = setupDataSource(QueryStepsDataSource);
       backendServer.fetch
-        .calledWith(requestMatching({ url: '/nitestmonitor/v2/query-steps', method: 'POST' }))
+        .calledWith(requestMatching({ url: queryStepsUrl, method: 'POST' }))
         .mockReturnValue(createFetchResponse(mockQueryStepsResponse));
       const query = buildQuery({
         refId: 'A',
@@ -435,7 +427,7 @@ describe('QueryStepsDataSource', () => {
       jest.spyOn(DataSourceBase.prototype, 'getWorkspaces').mockResolvedValue(mockWorkspaces);
       const [datastore, backendServer] = setupDataSource(QueryStepsDataSource);
       backendServer.fetch
-        .calledWith(requestMatching({ url: '/nitestmonitor/v2/query-steps', method: 'POST' }))
+        .calledWith(requestMatching({ url: queryStepsUrl, method: 'POST' }))
         .mockReturnValue(createFetchResponse(mockQueryStepsResponse));
       const query = buildQuery({
         refId: 'A',
@@ -494,7 +486,7 @@ describe('QueryStepsDataSource', () => {
           };
 
           backendServer.fetch
-            .calledWith(requestMatching({ url: '/nitestmonitor/v2/query-steps', method: 'POST' }))
+            .calledWith(requestMatching({ url: queryStepsUrl, method: 'POST' }))
             .mockReturnValue(createFetchResponse(mockQueryStepsMeasurementResponse));
         });
 
@@ -549,7 +541,7 @@ describe('QueryStepsDataSource', () => {
         };
 
         backendServer.fetch
-          .calledWith(requestMatching({ url: '/nitestmonitor/v2/query-steps', method: 'POST' }))
+          .calledWith(requestMatching({ url: queryStepsUrl, method: 'POST' }))
           .mockReturnValue(createFetchResponse(mockQueryStepsMeasurementResponse));
 
         const query = buildQuery({
@@ -587,7 +579,7 @@ describe('QueryStepsDataSource', () => {
         };
 
         backendServer.fetch
-          .calledWith(requestMatching({ url: '/nitestmonitor/v2/query-steps', method: 'POST' }))
+          .calledWith(requestMatching({ url: queryStepsUrl, method: 'POST' }))
           .mockReturnValue(createFetchResponse(mockQueryStepsMeasurementResponse));
 
         const query = buildQuery({
@@ -637,7 +629,7 @@ describe('QueryStepsDataSource', () => {
         };
 
         backendServer.fetch
-          .calledWith(requestMatching({ url: '/nitestmonitor/v2/query-steps', method: 'POST' }))
+          .calledWith(requestMatching({ url: queryStepsUrl, method: 'POST' }))
           .mockReturnValue(createFetchResponse(mockQueryStepsConditionsResponse));
       });
 
@@ -671,7 +663,7 @@ describe('QueryStepsDataSource', () => {
 
       test('should not create input or output columns when they are returned as empty arrays', async () => {
         backendServer.fetch
-          .calledWith(requestMatching({ url: '/nitestmonitor/v2/query-steps', method: 'POST' }))
+          .calledWith(requestMatching({ url: queryStepsUrl, method: 'POST' }))
           .mockReturnValue(
             createFetchResponse({
               steps: [
@@ -736,7 +728,7 @@ describe('QueryStepsDataSource', () => {
         };
 
         backendServer.fetch
-          .calledWith(requestMatching({ url: '/nitestmonitor/v2/query-steps', method: 'POST' }))
+          .calledWith(requestMatching({ url: queryStepsUrl, method: 'POST' }))
           .mockReturnValue(createFetchResponse(mockQueryStepsConditionsResponse));
       });
 
@@ -821,7 +813,7 @@ describe('QueryStepsDataSource', () => {
       };
 
       backendServer.fetch
-        .calledWith(requestMatching({ url: '/nitestmonitor/v2/query-steps', method: 'POST' }))
+        .calledWith(requestMatching({ url: queryStepsUrl, method: 'POST' }))
         .mockReturnValue(createFetchResponse(mockQueryStepsConditionsResponse));
 
       const query = buildQuery({
@@ -915,7 +907,7 @@ describe('QueryStepsDataSource', () => {
     });
 
     test('should handle null and undefined properties', async () => {
-      backendServer.fetch.calledWith(requestMatching({ url: '/nitestmonitor/v2/query-steps' })).mockReturnValue(
+      backendServer.fetch.calledWith(requestMatching({ url: queryStepsUrl })).mockReturnValue(
         createFetchResponse({
           steps: [
             {
@@ -952,7 +944,7 @@ describe('QueryStepsDataSource', () => {
         }),
       ];
       backendServer.fetch
-        .calledWith(requestMatching({ url: '/nitestmonitor/v2/query-steps', method: 'POST' }))
+        .calledWith(requestMatching({ url: queryStepsUrl, method: 'POST' }))
         .mockImplementationOnce(() => mockResponses[0])
         .mockImplementationOnce(() => mockResponses[1]);
       const query = buildQuery({
@@ -968,7 +960,7 @@ describe('QueryStepsDataSource', () => {
       const fields = response.data[0].fields as Field[];
       expect(fields).toEqual([{ name: 'A', values: [5000] }]);
       const callsToQuerySteps = backendServer.fetch.mock.calls.filter(
-        ([request]) => request.url === '/nitestmonitor/v2/query-steps'
+        ([request]) => request.url === queryStepsUrl
       );
       expect(callsToQuerySteps).toHaveLength(1);
     });
@@ -1339,10 +1331,10 @@ describe('QueryStepsDataSource', () => {
         {path: '    Path with spaces    '},
       ]
       backendServer.fetch
-      .calledWith(requestMatching({ url: '/nitestmonitor/v2/query-paths', method: 'POST' }))
+      .calledWith(requestMatching({ url: queryStepsUrl, method: 'POST' }))
       .mockReturnValue(
         createFetchResponse({
-          paths: mockPaths,
+          steps: mockPaths,
           continuationToken: null,
         })
       );
@@ -1375,10 +1367,10 @@ describe('QueryStepsDataSource', () => {
 
     it('should return an empty array when no step paths are returned', async () => {
       backendServer.fetch
-      .calledWith(requestMatching({ url: '/nitestmonitor/v2/query-paths', method: 'POST' }))
+      .calledWith(requestMatching({ url: queryStepsUrl, method: 'POST' }))
       .mockReturnValue(
         createFetchResponse({
-          paths: [],
+          steps: [],
           continuationToken: null,
         })
       );
@@ -1391,7 +1383,7 @@ describe('QueryStepsDataSource', () => {
     it('should make a single request when take is less than MAX_PATH_TAKE_PER_REQUEST', async () => {
       const mockResponses = [
         createFetchResponse({
-          paths: mockPaths.slice(0, 100),
+          steps: mockPaths.slice(0, 100),
           continuationToken: null,
         }),
       ];
@@ -1404,7 +1396,7 @@ describe('QueryStepsDataSource', () => {
       expect(backendServer.fetch).toHaveBeenNthCalledWith(
         1,
         expect.objectContaining({
-          url: '/nitestmonitor/v2/query-paths',
+          url: queryStepsUrl,
           data: expect.objectContaining({ take: 100, continuationToken: undefined }),
         })
       );
@@ -1413,11 +1405,11 @@ describe('QueryStepsDataSource', () => {
     test('should batch requests when total number of paths matching the filter is less than requested take', async () => {
       const mockResponses = [
         createFetchResponse({
-          paths: mockPaths,
+          steps: mockPaths,
           continuationToken: 'token1',
         }),
         createFetchResponse({
-          paths: mockPaths.slice(0, 500),
+          steps: mockPaths.slice(0, 500),
           continuationToken: null,
         }),
       ];
@@ -1430,14 +1422,14 @@ describe('QueryStepsDataSource', () => {
       expect(backendServer.fetch).toHaveBeenNthCalledWith(
         1,
         expect.objectContaining({
-          url: '/nitestmonitor/v2/query-paths',
-          data: expect.objectContaining({ take: 1000, continuationToken: undefined }),
+          url: queryStepsUrl,
+          data: expect.objectContaining({ take: 500, continuationToken: undefined }),
         })
       );
       expect(backendServer.fetch).toHaveBeenNthCalledWith(
         2,
         expect.objectContaining({
-          url: '/nitestmonitor/v2/query-paths',
+          url: queryStepsUrl,
           data: expect.objectContaining({ take: 1000, continuationToken: 'token1' }),
         })
       );
@@ -1449,19 +1441,19 @@ describe('QueryStepsDataSource', () => {
 
       const mockResponses = [
         createFetchResponse({
-          paths: mockPaths,
+          steps: mockPaths,
           continuationToken: 'token1',
         }),
         createFetchResponse({
-          paths: mockPaths,
+          steps: mockPaths,
           continuationToken: 'token2',
         }),
         createFetchResponse({
-          paths: mockPaths,
+          steps: mockPaths,
           continuationToken: 'token3',
         }),
         createFetchResponse({
-          paths: mockPaths,
+          steps: mockPaths,
           continuationToken: null,
         }),
       ];
@@ -1483,28 +1475,28 @@ describe('QueryStepsDataSource', () => {
       expect(fetchSpy).toHaveBeenNthCalledWith(
         1,
         expect.objectContaining({
-          url: '/nitestmonitor/v2/query-paths',
-          data: expect.objectContaining({ take: 1000, continuationToken: undefined }),
+          url: queryStepsUrl,
+          data: expect.objectContaining({ take: 500, continuationToken: undefined }),
         })
       );
       expect(fetchSpy).toHaveBeenNthCalledWith(
         2,
         expect.objectContaining({
-          url: '/nitestmonitor/v2/query-paths',
+          url: queryStepsUrl,
           data: expect.objectContaining({ take: 1000, continuationToken: 'token1' }),
         })
       );
       expect(fetchSpy).toHaveBeenNthCalledWith(
         3,
         expect.objectContaining({
-          url: '/nitestmonitor/v2/query-paths',
+          url: queryStepsUrl,
           data: expect.objectContaining({ take: 1000, continuationToken: 'token2' }),
         })
       );
       expect(fetchSpy).toHaveBeenNthCalledWith(
         4,
         expect.objectContaining({
-          url: '/nitestmonitor/v2/query-paths',
+          url: queryStepsUrl,
           data: expect.objectContaining({ take: 1000, continuationToken: 'token3' }),
         })
       );
@@ -1513,7 +1505,7 @@ describe('QueryStepsDataSource', () => {
     test('should stop fetching when continuationToken is null', async () => {
       const mockResponses = [
         createFetchResponse({
-          paths: mockPaths,
+          steps: mockPaths,
           continuationToken: null,
         }),
       ];
@@ -1525,8 +1517,8 @@ describe('QueryStepsDataSource', () => {
       expect(backendServer.fetch).toHaveBeenCalledTimes(1);
       expect(backendServer.fetch).toHaveBeenCalledWith(
         expect.objectContaining({
-          url: '/nitestmonitor/v2/query-paths',
-          data: expect.objectContaining({ take: 1000, continuationToken: undefined }),
+          url: queryStepsUrl,
+          data: expect.objectContaining({ take: 500, continuationToken: undefined }),
         })
       );
     });
@@ -1535,14 +1527,14 @@ describe('QueryStepsDataSource', () => {
       backendServer.fetch
         .mockImplementationOnce(() =>
           createFetchResponse({
-            paths: mockPaths,
+            steps: mockPaths,
             continuationToken: 'token1',
           })
         )
         .mockImplementationOnce(() => createFetchError(400)) //Error
         .mockImplementationOnce(() =>
           createFetchResponse({
-            paths: mockPaths,
+            steps: mockPaths,
             continuationToken: 'token2',
           })
         );
@@ -1550,7 +1542,7 @@ describe('QueryStepsDataSource', () => {
       const batchPromise = datastore.queryStepPathInBatches('filter', undefined, 3000, true);
 
       await expect(batchPromise).rejects.toThrow(
-        'Request to url "/nitestmonitor/v2/query-paths" failed with status code: 400. Error message: "Error"'
+        'The query failed due to the following error: (status 400) \"Error\".'
       );
       expect(backendServer.fetch).toHaveBeenCalledTimes(2);
     });
@@ -1567,19 +1559,19 @@ describe('QueryStepsDataSource', () => {
 
       const mockResponses = [
         createFetchResponse({
-          paths: mockPaths,
+          steps: mockPaths,
           continuationToken: 'token1',
         }),
         createFetchResponse({
-          paths: mockPaths,
+          steps: mockPaths,
           continuationToken: 'token2',
         }),
         createFetchResponse({
-          paths: mockPaths,
+          steps: mockPaths,
           continuationToken: 'token3',
         }),
         createFetchResponse({
-          paths: mockPaths,
+          steps: mockPaths,
           continuationToken: null,
         }),
       ];
@@ -1617,14 +1609,14 @@ describe('QueryStepsDataSource', () => {
     });
 
     it('should return unique step paths', async () => {
-      backendServer.fetch
-        .calledWith(requestMatching({ url: '/nitestmonitor/v2/query-paths', method: 'POST' }))
-        .mockReturnValue(
-          createFetchResponse({
-            paths: [{ path: 'path1' }, { path: 'path2' }, { path: 'path1' }, { path: 'path3' }, { path: 'path2' }],
-            continuationToken: null,
-          })
-        );
+      backendServer.fetch.calledWith(requestMatching({ url: queryStepsUrl })).mockReturnValue(
+        createFetchResponse({
+          steps: [
+            { path: 'path1' }, { path: 'path2' }, { path: 'path1' }, { path: 'path3' }, { path: 'path2' }
+          ],
+          continuationToken: null,
+        } as unknown as QueryStepsResponse)
+      );
 
       const result = await datastore.getStepPaths('ProgramName = "Test"');
 
@@ -1644,7 +1636,7 @@ describe('QueryStepsDataSource', () => {
       expect(spy).not.toHaveBeenCalled();
     });
 
-    it('should handle error in query-paths API when loading step path', async () => {
+    it('should handle error in query-steps API when loading step path', async () => {
       const error = new Error('API failed');
       jest.spyOn(datastore as any, 'loadStepPaths').mockRejectedValue(error);
 
@@ -1657,7 +1649,7 @@ describe('QueryStepsDataSource', () => {
       );
     });
 
-    it('should handle 504 errors in query-paths API when loading step path', async () => {
+    it('should handle 504 errors in query-steps API when loading step path', async () => {
       const error = new Error(
         `API failed Error message: status code: 504 ${JSON.stringify({ message: 'Detailed error message' })}`
       );
@@ -1699,18 +1691,31 @@ describe('QueryStepsDataSource', () => {
       );
     });
 
-    it('should contain error details when query-path error contains additional information', async () => {
+    it('should contain error details when query-steps error contains additional information', async () => {
       const error = new Error(
         `API failed Error message: ${JSON.stringify({ message: 'Detailed error message', statusCode: 500 })}`
       );
       jest.spyOn(datastore as any, 'queryResultsValues').mockResolvedValue(['name1', 'name2']);
-      jest.spyOn(datastore as any, 'queryStepPaths').mockRejectedValue(error);
+      jest.spyOn(datastore as any, 'queryStepsInBatches').mockRejectedValue(error);
 
       await datastore.getStepPaths('ProgramName = "Test"');
 
       expect(datastore.errorTitle).toBe('Warning during step paths value query');
       expect(datastore.errorDescription).toContain(
         'Some values may not be available in the query builder lookups due to the following error:Detailed error message.'
+      );
+    });
+
+    it('should include result filter in query when resultsQuery is provided', async () => {
+      const spy = jest.spyOn(datastore as any, 'queryStepPathInBatches');
+      jest.spyOn(datastore as any, 'queryResultsValues').mockResolvedValue(['name1', 'name2']);
+
+      await datastore.getStepPaths('PartNumber = "123"');
+
+      expect(spy).toHaveBeenCalledWith(
+        '(PartNumber = \"123\") && (ProgramName = \"name1\" || ProgramName = \"name2\")',
+        ['path'],
+        1000
       );
     });
   });
@@ -1727,7 +1732,7 @@ describe('QueryStepsDataSource', () => {
 
       expect(backendServer.fetch).toHaveBeenCalledWith(
         expect.objectContaining({
-          url: '/nitestmonitor/v2/query-steps',
+          url: queryStepsUrl,
           data: expect.objectContaining({
             resultFilter: 'ProgramName = "name1"',
             filter: 'stepType = "Type1"',
@@ -1746,7 +1751,7 @@ describe('QueryStepsDataSource', () => {
 
       expect(backendServer.fetch).toHaveBeenCalledWith(
         expect.objectContaining({
-          url: '/nitestmonitor/v2/query-steps',
+          url: queryStepsUrl,
           data: expect.objectContaining({
             resultFilter: '(ProgramName = "name1" || ProgramName = "name2")',
           }),
@@ -1766,7 +1771,7 @@ describe('QueryStepsDataSource', () => {
 
       expect(backendServer.fetch).toHaveBeenCalledWith(
         expect.objectContaining({
-          url: '/nitestmonitor/v2/query-steps',
+          url: queryStepsUrl,
           data: expect.objectContaining({
             resultFilter: 'UpdatedAt = "2025-01-01T00:00:00.000Z"',
           }),
@@ -1787,7 +1792,7 @@ describe('QueryStepsDataSource', () => {
 
       expect(backendServer.fetch).toHaveBeenCalledWith(
         expect.objectContaining({
-          url: '/nitestmonitor/v2/query-steps',
+          url: queryStepsUrl,
           data: expect.objectContaining({
             resultFilter: '(PartNumber = "123" || Keywords != "456") && HostName contains "Test"',
             filter: '(stepType = "123" || keywords != "456") && name contains "Test"',
@@ -1892,7 +1897,7 @@ describe('QueryStepsDataSource', () => {
 
       expect(backendServer.fetch).toHaveBeenCalledWith(
         expect.objectContaining({
-          url: '/nitestmonitor/v2/query-steps',
+          url: queryStepsUrl,
           data: expect.objectContaining({
             orderBy: 'STARTED_AT',
             descending: false,
@@ -1951,7 +1956,7 @@ describe('QueryStepsDataSource', () => {
     it('should return undefined if API throws error', async () => {
       const error = new Error('API failed');
       backendServer.fetch
-        .calledWith(requestMatching({ url: '/nitestmonitor/v2/query-steps', method: 'POST' }))
+        .calledWith(requestMatching({ url: queryStepsUrl, method: 'POST' }))
         .mockImplementationOnce(() => {
           throw error;
         });
@@ -2004,7 +2009,7 @@ describe('QueryStepsDataSource', () => {
 
       expect(backendServer.fetch).toHaveBeenCalledWith(
         expect.objectContaining({
-          url: '/nitestmonitor/v2/query-steps',
+          url: queryStepsUrl,
           data: expect.objectContaining({
             resultFilter: 'ProgramName = "name1"',
           }),